add test_reports.py
