+++ tests/test_utils.py
@@ -1,9 +1,12 @@
+from datetime import datetime
 from unittest.mock import patch
 
-from src.reader import reader_excel_file
+import pytest
+
+from src.utils import reader_excel_file, greeting, get_start_and_end_date
 
 
-@patch("src.reader.pd.read_excel")
+@patch("src.utils.pd.read_excel")
 def test_reader_excel_file(mock_read_excel):
     """Тест при excel файле c данными"""
     mock_read_excel.return_value.to_dict.return_value = """[
@@ -100,3 +103,33 @@ def test_reader_excel_file_file_invalid_file_format(txt_file):
     """Тест, если файл с неправильным форматом"""
     result = reader_excel_file(txt_file)
     assert result == "Файл пустой или имеет неподдерживаемый формат."
+
+
+@pytest.mark.parametrize(
+    "test_time, expected",
+    [
+        ("2025-03-05 08:10:15", "Доброе утро!"),
+        ("2025-03-05 14:10:15", "Добрый день!"),
+        ("2025-03-05 18:10:15", "Добрый вечер!"),
+        ("2025-03-05 02:10:15", "Доброй ночи!"),
+    ],
+)
+def test_greeting(test_time, expected):
+    """Тест на правильное приветствие"""
+    mock_time = datetime.strptime(test_time, "%Y-%m-%d %H:%M:%S")
+    with patch("src.utils.datetime") as mock_datetime:
+        mock_datetime.now.return_value = mock_time
+        assert greeting() == expected
+
+
+def test_get_start_and_end_date():
+    """Проверка функцию на правильную выдачу диапазона"""
+    result = get_start_and_end_date("2018-03-05 08:10:15")
+    assert result == (datetime(2018, 3, 1, 0, 0, 0), datetime(2018, 3, 5, 8, 10, 15))
+
+
+def test_get_start_and_end_date_error():
+    """Тест на обработку функции при неправильном формате даты"""
+    result = get_start_and_end_date("03-05-2018 08:10:15")
+    assert result == "Произошла ошибка time data '03-05-2018 08:10:15' does not match format '%Y-%m-%d %H:%M:%S'"
+
