+++ tests/test_views.py
@@ -3,37 +3,8 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from src.views import (get_currency_rates, get_data_of_cards, get_start_and_end_date, get_stock_prices,
-                       get_top_list_transction, greeting)
-
-
-@pytest.mark.parametrize(
-    "test_time, expected",
-    [
-        ("2025-03-05 08:10:15", "Доброе утро!"),
-        ("2025-03-05 14:10:15", "Добрый день!"),
-        ("2025-03-05 18:10:15", "Добрый вечер!"),
-        ("2025-03-05 02:10:15", "Доброй ночи!"),
-    ],
-)
-def test_greeting(test_time, expected):
-    """Тест на правильное приветствие"""
-    mock_time = datetime.strptime(test_time, "%Y-%m-%d %H:%M:%S")
-    with patch("src.views.datetime") as mock_datetime:
-        mock_datetime.now.return_value = mock_time
-        assert greeting() == expected
-
-
-def test_get_start_and_end_date():
-    """Проверка функцию на правильную выдачу диапазона"""
-    result = get_start_and_end_date("2018-03-05 08:10:15")
-    assert result == (datetime(2018, 3, 1, 0, 0, 0), datetime(2018, 3, 5, 8, 10, 15))
-
-
-def test_get_start_and_end_date_error():
-    """Тест на обработку функции при неправильном формате даты"""
-    result = get_start_and_end_date("03-05-2018 08:10:15")
-    assert result == "Произошла ошибка time data '03-05-2018 08:10:15' does not match format '%Y-%m-%d %H:%M:%S'"
+from src.views import (get_currency_rates, get_data_of_cards, get_stock_prices,
+                       get_top_list_transction)
 
 
 @pytest.mark.parametrize(
